// src/types/settings.ts
export interface Settings {
  downloadPath: string;
  subtitlePath?: string;
  maxDownloadSpeed: string;
  maxUploadSpeed: string;
  maxActiveTorrents: string;
  preferredGroup: string;
  autoAddNewTorrents: boolean;
  autoStart: boolean;
  darkMode: boolean;
  notifications: boolean;
  videoQuality: string;
  subtitle: boolean;
  subtitleLanguage: string;
}